Fixes after review
- Changed redux structure
- Refactor component functions to use ES6 shorthand syntax without "return" keyword
