+++ src/components/App.js
@@ -5,53 +5,51 @@ import { connect } from 'react-redux';
 
 import { ImageList, SearchForm } from '.';
 
-const App = ({ imageData, loading }) => {
-  return (
-    <Container className="mb-5">
-      <h2 className="text-center my-4">Walty - graphics for all your needs</h2>
+const App = ({ imageData, loading }) => (
+  <Container className="mb-5">
+    <h2 className="text-center my-4">Walty - graphics for all your needs</h2>
 
-      <Row className="mb-4">
-        <Col>
-          <SearchForm />
-        </Col>
-      </Row>
+    <Row className="mb-4">
+      <Col>
+        <SearchForm />
+      </Col>
+    </Row>
 
-      <Row className="mb-5">
-        <Col>
-          {loading && <p className="text-center mt-2">Loading...</p>}
+    <Row className="mb-5">
+      <Col>
+        {loading && <p className="text-center mt-2">Loading...</p>}
 
-          {imageData.total > 0 && (
-            <p className="text-right mt-2">
-              Results: {imageData.totalHits} of {imageData.total}
-            </p>
-          )}
+        {imageData.total > 0 && (
+          <p className="text-right mt-2">
+            Results: {imageData.totalHits} of {imageData.total}
+          </p>
+        )}
 
-          <ImageList images={imageData.images} />
-        </Col>
-      </Row>
+        <ImageList images={imageData.images} />
+      </Col>
+    </Row>
 
-      <Row>
-        <Col>
-          <div className="d-flex align-items-center justify-content-center">
-            Powered by{' '}
-            <a
-              href="http://pixabay.com/"
-              rel="noreferrer"
-              target="_blank"
-              className="ml-2"
-            >
-              <img
-                src="https://pixabay.com/static/img/logo.svg"
-                alt="Pixabay"
-                width="150"
-              />
-            </a>
-          </div>
-        </Col>
-      </Row>
-    </Container>
-  );
-};
+    <Row>
+      <Col>
+        <div className="d-flex align-items-center justify-content-center">
+          Powered by{' '}
+          <a
+            href="http://pixabay.com/"
+            rel="noreferrer"
+            target="_blank"
+            className="ml-2"
+          >
+            <img
+              src="https://pixabay.com/static/img/logo.svg"
+              alt="Pixabay"
+              width="150"
+            />
+          </a>
+        </div>
+      </Col>
+    </Row>
+  </Container>
+);
 
 App.propTypes = {
   imageData: PropTypes.shape({
